// src/app/models/Figure.ts
export interface Figure {

    journalId: string;
    figureId: string,
    fileName: string;
    fileType: string;
    iconType: string;
    data: any;
    imageURL: any;
    /*
    Word document: description
    Excel document: assessment
    PDF document: library_books
    image: insert_photo
    */
}